backend/requirements.txt
flask
tensorflow
pillow
numpy
pip freeze > requirements.txt
flask-cors
scikit-learn
joblib
